# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LBX Task Documentation",
        "description": "Implement Employee CRUD With Excel Sheet Import",
        "contact": {
            "url": "https://www.linkedin.com/in/abdallah-zaghloul",
            "email": "3bdallahzaghloul@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/excelSheet/{id}": {
            "get": {
                "tags": [
                    "ExcelSheet"
                ],
                "summary": "Show ExcelSheet",
                "description": "Return ExcelSheet",
                "operationId": "showExcelSheet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "excelSheetID as UUID like 9a4827a1-37ca-4222-b91b-3b551ba74d28",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "Successfull operation."
                                        },
                                        "data": {
                                            "properties": {
                                                "excel_sheet": {
                                                    "type": "object",
                                                    "default": {
                                                        "id": "9a48014d-fcc9-44ac-a13d-c9e72645f284",
                                                        "path": "ExcelSheet/9a48014d-fcc9-44ac-a13d-c9e72645f284.csv",
                                                        "status": "Processing",
                                                        "errors": null,
                                                        "created_at": "03/10/2023 01:41:43 PM",
                                                        "updated_at": "03/10/2023 01:41:43 PM",
                                                        "url": "https://large-file-upload-laravel.s3.eu-central-1.amazonaws.com/ExcelSheet/9a48014d-fcc9-44ac-a13d-c9e72645f284.csv",
                                                        "updates_url": "http://localhost:8000/api/excelSheet/9a48014d-fcc9-44ac-a13d-c9e72645f284"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "default": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "Not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Unavailable Server",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "default": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "default": "Sorry something went wrong ... please try again later."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}